change get hourly data and try to get again when fail
