+++ app/src/main/java/com/example/airquality/services/sensors/APIDeviceService.kt
@@ -1,5 +1,6 @@
 package com.example.airquality.services.sensors
 
+import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -7,6 +8,7 @@ import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.PUT
+import java.util.concurrent.TimeUnit
 
 interface APIDeviceService {
     @GET("getLatest")
